Split custom metadata specs
